First Commit
This is a better organized recreation (?) of my pygame sokoban prototype.  The undo stack works perfectly for movement, creation, and deletion (though no game mechanics actually create or delete any objects, yet).  Right now you can push a block around, and there's a wall, and that's about it.  Everything is still in one file at the moment, which is not great.